+++ .clang-format
@@ -3,9 +3,9 @@
 ---
 Language: Cpp
 BasedOnStyle: Chromium
-IndentWidth: 4
+IndentWidth: 2
 
-AccessModifierOffset: -4
+AccessModifierOffset: -2
 AlignAfterOpenBracket: AlwaysBreak
 AlignConsecutiveAssignments: Consecutive
 AlignEscapedNewlines: Left
@@ -47,8 +47,8 @@ BreakStringLiterals: false
 BreakInheritanceList: AfterColon
 BreakConstructorInitializers: AfterColon
 
-ContinuationIndentWidth: 4
-ConstructorInitializerIndentWidth: 4
+ContinuationIndentWidth: 2
+ConstructorInitializerIndentWidth: 2
 Cpp11BracedListStyle: true
 IncludeCategories:
   # POSIX
