--- platform_tilt_brute_forcer/platform_tilt_brute_forcer.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "platform_tilt_brute_forcer", "platform_tilt_brute_forcer\platform_tilt_brute_forcer.vcxproj", "{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Debug|x64.ActiveCfg = Debug|x64
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Debug|x64.Build.0 = Debug|x64
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Debug|x86.ActiveCfg = Debug|Win32
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Debug|x86.Build.0 = Debug|Win32
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Release|x64.ActiveCfg = Release|x64
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Release|x64.Build.0 = Release|x64
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Release|x86.ActiveCfg = Release|Win32
-		{CF8C2F81-2CE9-48F2-83CB-F740C87838E9}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {95188BDC-6C35-43FE-832D-6A8B68970067}
-	EndGlobalSection
-EndGlobal
