+++ src/Trig.cpp
