+++ src/Trig.hpp
