+++ src/main.cpp
@@ -1,0 +1,1 @@
+int main() {}
