+++ src/solvers.cl
