+++ src/solvers.h.cl
