+++ src/vmath.cpp
